explicitly commit changes
this change only affects the internal logic: Btree.insert, Btree.delete
and Btree.mark_deleted now all do not commit the change: Btree.commit
has to be explicitly called to update the header.

This change has no impact on the public API, but makes the code simpler
(Btree.insert, Btree.delete and Btree.mark_deleted do not need a
commit_header optional argument anymore), and also allows an
optimization of the catch-up process at the end of a compaction.