improve get_multi interface (breaking change)
Before, get_multi/3 was returning a list of value corresponding to the
given keys, or the default value when the key was not in the database.

This API had a number of drawbacks, like the difficulty to distinguish
between a missing key from a key associated to the default value, or the
inconvenience of matching keys to values. Also, it is not consistent
with the argument of the function passed to get_and_update_multi.

The new version solves these problems (at the cost of breaking backward
compatibility) by returning a map of keys to values fro the selected
keys. When a key is not present in the database, it is omitted from the
result map.

Example:

    # Assuming that the database contains a: 1, b: 2, c: nil

    # Before:
    CubDB.get_multi(db, [:a, :b, :c, :x])
    # => [1, 2, nil, nil]

    # Now:
    CubDB.get_multi(db, [:a, :b, :c, :x])
    # => %{a: 1, b: 2, c: nil}