raise if the same file is used by multiple stores
this will help avoiding the mistake of starting multiple CubDB processes
on the same data directory. Multiple CubDB processes using the same data
file can cause data corruption, so we prevent this and alert the user
when possible.

Note that this mechanism does not yet prevent a user to mistakenly start
two CubDB processes using the same data directory on multiple VMs.