better timeout logic for read and write operations
Rather than relying on the GenServer.call timeout, that exits the caller
but does not free up resources on the callee, this commit enforces the
timeout also on the callee side when performing reads.

Also `get_and_update_multi/4` now enforces a timeout on the callee side.

This commit also cleans up the public API around timeouts by introducing
a breaking change, as both `select/2` and `get_and_update_multi/4` now
accept the timeout as an option instead of a separate argument. This
allows to introduce more options in the future, without having to
introduce more breaking changes.

The reader check in and check out is also refactored to be substantially
simpler.