fix file descriptor leak
when cleaning up old files, the corresponding Store should also be
stopped, releasing the file descriptor. This step was missing, leading
to a file descriptor leak that would manifest after many compactions.

Now the CubDB process keeps a list of old B-trees, and stops them
(closing the relative Store) upon cleanup, when it ensured that those
b-trees and their files are not in use anymore.