Take zero cost immutable snapshots of the database and read from them
A snapshot can be obtained with `CubDB.snapshot/2`, and represents the
state of the database in the instant it was taken. Taking a snapshot is
almost zero cost: nothing needs to be copied or written to disk.

After taking a snapshot, it is possible to perform read operations on
it, with the usual `get`, `get_multi`, `fetch`, `select`, `has_key?`.

The only real cost of a snapshot is that some cleanup operations after a
compaction have to wait while a snapshot is in use. This is why
snapshots have by default a timeout, after which they cannot be used
anymore. It is possible to specify a timeout of `:infinity`, but then
one has to manually call `CubDB.release_snapshot/1` to release
resources.