Improve Reader module and change select/2 return signature
Now that `select/2`, like other operations, is executed on the caller
process, and now that the `:timeout` option is not available anymore,
there is no need to return a `{:ok, result}` tuple. Avoiding to catch
arbitrary exceptions to return `{:error, exception}` is also a better
API.

The `Reader` module now uses named functions instead of `perform/2`, as
it is not executed in a `Task` anymore.