Rename CubDN.select_stream/2 to CubDB.select/2
Replacing the old `select/2` with a new version that does not accept
`:reduce` and `:pipe` options, but instead returns a lazy stream that
can be used with functions in `Stream` and `Enum`.

While this represents a breaking change, it makes the API more
convenient, and leverages standard functions in the core library instead
of custom options.